Move converter functions from types into XAccountController
